Создал новый класс для клиента ConsoleClient, вынес функционал клиента в него
